feat(Profile): split cards into components and add language selector to therapist profile
