+++ frontend/src/pages/PatientProfile.tsx
@@ -1,70 +1,29 @@
-import { useEffect, useState } from 'react';
+import { useEffect } from 'react';
 import Layout from '@/components/Layout';
 import PageHeader from '@/components/PageHeader';
 import Section from '@/components/Section';
 import { useTranslation } from 'react-i18next';
-import { Switch } from '@/components/ui/switch';
 import { Button } from '@/components/ui/button';
 import { Badge } from '@/components/ui/badge';
-import { Skeleton } from '@/components/ui/skeleton';
-import FitbitConnectButton from '@/components/PatientPage/FitbitStatus';
-import { useNotifications } from '@/hooks/useNotifications';
+import Card from '@/components/Card';
 import authStore from '@/stores/authStore';
-import flagDe from '@/assets/flags/de.png';
-import flagFr from '@/assets/flags/fr.png';
-import flagEn from '@/assets/flags/gb.png';
-import flagIt from '@/assets/flags/it.png';
-import flagPt from '@/assets/flags/pt.png';
-import flagNl from '@/assets/flags/be.png';
 import { observer } from 'mobx-react-lite';
 import { Link, useNavigate } from 'react-router-dom';
-import {
-  Select,
-  SelectContent,
-  SelectGroup,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '@/components/ui/select';
 import LogoutFill from '@/assets/icons/logout-fill.svg?react';
 import Mail from '@/assets/icons/contact/mail.svg?react';
 import Phone from '@/assets/icons/contact/phone.svg?react';
 import config from '@/config/config.json';
 import { patientFitbitStore } from '@/stores/patientFitbitStore';
-import Card from '@/components/Card';
+import LanguageSelectorCard from '@/components/UserProfile/LanguageSelectorCard';
+import NotificationsCard from '@/components/UserProfile/NotificationsCard';
+import FitbitCard from '@/components/UserProfile/FitbitCard';
 
 const PatientProfile: React.FC = observer(() => {
-  const { t, i18n } = useTranslation();
+  const { t } = useTranslation();
   const navigate = useNavigate();
-  const { enabled, permission, supportsPeriodicSync, toggleNotifications } = useNotifications();
 
   const patientId = localStorage.getItem('id') || authStore.id || '';
   const displayName = authStore.firstName || t('Profile');
-
-  const getInitialLang = () =>
-    localStorage.getItem('i18nextLng')?.slice(0, 2) || i18n.language?.slice(0, 2) || 'en';
-
-  const [currentLanguage, setCurrentLanguage] = useState(getInitialLang);
-
-  const languages = ['de', 'fr', 'en', 'it', 'pt', 'nl'] as const;
-
-  const flagMap: Record<string, string> = {
-    en: flagEn,
-    de: flagDe,
-    fr: flagFr,
-    it: flagIt,
-    pt: flagPt,
-    nl: flagNl,
-  };
-
-  const languageNames: Record<string, string> = {
-    en: 'English',
-    de: 'Deutsch',
-    fr: 'Français',
-    it: 'Italiano',
-    pt: 'Português',
-    nl: 'Nederlands',
-  };
 
   const partnerLogos = [
     {
@@ -83,13 +42,6 @@ const PatientProfile: React.FC = observer(() => {
       className: 'w-[160px]',
     },
   ];
-
-  const lang = currentLanguage.slice(0, 2);
-
-  const handleLanguageChange = (l: string) => {
-    i18n.changeLanguage(l);
-    setCurrentLanguage(l);
-  };
 
   const handleLogout = async () => {
     await authStore.logout();
@@ -127,72 +79,9 @@ const PatientProfile: React.FC = observer(() => {
         <Section>
           <div className="p-2 pl-4 font-medium text-lg text-zinc-500">{t('Settings')}</div>
 
-          <Card className="flex flex-col gap-1">
-            <div className="text-sm font-medium text-zinc-500">{t('Language')}</div>
-            <Select onValueChange={(value) => handleLanguageChange(value)} defaultValue={lang}>
-              <SelectTrigger className="bg-white border-white shadow-none p-0">
-                <SelectValue placeholder={t('Select language')} />
-              </SelectTrigger>
-              <SelectContent className="bg-zinc-50 rounded-3xl p-1">
-                <SelectGroup>
-                  {languages.map((l) => (
-                    <SelectItem key={l} value={l}>
-                      <span className="font-bold text-lg leading-6 text-zinc-800">
-                        {languageNames[l]}
-                      </span>
-                      <img src={flagMap[l]} className="h-4 w-4 rounded-full ml-1 -mt-1" />
-                    </SelectItem>
-                  ))}
-                </SelectGroup>
-              </SelectContent>
-            </Select>
-          </Card>
-
-          <Card className="flex flex-col gap-1">
-            <div className="text-sm font-medium text-zinc-500">{t('Notifications')}</div>
-            <div className="flex items-center justify-between">
-              <div className="flex flex-col">
-                <div className="font-bold text-lg leading-6 text-zinc-800">
-                  {t('Receive reminders')}
-                </div>
-                {permission === 'denied' && (
-                  <div className="text-nok text-xs">
-                    {t('Notification permission denied. Please enable in browser settings.')}
-                  </div>
-                )}
-                {!supportsPeriodicSync && (
-                  <div className="text-amber-600 text-xs">
-                    {t('Background notifications not supported in this browser.')}
-                  </div>
-                )}
-              </div>
-              <Switch checked={enabled} onCheckedChange={toggleNotifications} />
-            </div>
-          </Card>
-
-          <div
-            className={`p-4 rounded-3xl flex gap-1 justify-between items-center ${patientFitbitStore.connected === false ? 'bg-zinc-100' : 'border border-accent'}`}
-          >
-            {patientFitbitStore.connected === true ? (
-              <div className="flex flex-col">
-                <div className="font-medium text-sm text-zinc-500">{t('Fitness Tracker')}</div>
-                <div className="font-bold text-lg text-zinc-800">{t('Fitbit Connected')}</div>
-              </div>
-            ) : patientFitbitStore.connected === false ? (
-              <>
-                <div className="flex flex-col">
-                  <div className="font-bold text-lg text-zinc-800">{t('Fitbit')}</div>
-                  <div className="font-medium text-sm text-zinc-500">{t('Fitness Tracker')}</div>
-                </div>
-                <FitbitConnectButton />
-              </>
-            ) : (
-              <div className="flex flex-col gap-2">
-                <Skeleton className="h-4 w-20" />
-                <Skeleton className="h-6 w-32" />
-              </div>
-            )}
-          </div>
+          <LanguageSelectorCard />
+          <NotificationsCard />
+          <FitbitCard />
         </Section>
 
         <Section>
